Edits and minor tweaks
Fixed having to have a hidden image to get our <Map> to show. All we need to do
is place <Map> inside a div, so that the image will fill to the dimensions of
div. Then, when you place the parent div in the right column of ManoaMap.jsx,
it will appear normal.